Fix crash in halfawake mode(GPU is off: 0Hz, but CPU is still working)
[!] The console still wakes itself even if system-config.ini is applied (once per ~13min10sec). Further investigation is required!

+++ Source/sys-clk-OC/common/include/sysclk/clocks.h
@@ -53,8 +53,8 @@ typedef struct
 typedef struct
 {
     union {
-        uint32_t mhz[SysClkProfile_EnumMax * SysClkModule_EnumMax];
-        uint32_t mhzMap[SysClkProfile_EnumMax][SysClkModule_EnumMax];
+        uint32_t mhz[(int)SysClkProfile_EnumMax * (int)SysClkModule_EnumMax];
+        uint32_t mhzMap[(int)SysClkProfile_EnumMax][(int)SysClkModule_EnumMax];
     };
 } SysClkTitleProfileList;
